try float 8

+++ oneat/NEATUtils/MovieCreator.py
@@ -563,13 +563,13 @@ yolo_v1 = True, yolo_v2 = False,  tshift  = 0, normalizeimage = True):
                       if name == Segname:
                           
                           
-                         image = daskread(fname)[0].astype('float16')
+                         image = daskread(fname)[0].astype('float8')
                          print('raw', image.shape)
                             
                          segimage = daskread(Segfname)[0].astype('float16')
                          print('seg',segimage.shape)
                          if normalizeimage:
-                            image = normalizeFloatZeroOne( image.compute().astype('float16'),1,99.8)
+                            image = normalizeFloatZeroOne( image.compute(),1,99.8)
                          print('image normalized')
                          for csvfname in filesCsv:
                                  count = 0
